Fix JVM connection. New problem with PQTree
Either an error in the java program, or it occurs at the stage of converting types to jpype.
It happens when checking the graph for planarity `.is Planar(graph)`.

Moreover, after the algorithm works, the library does not provide for calculating the coordinates of vertices on the plane.
There are a lot of TODO's in the library code, which indicates the inferiority of the library => the decision was made to abandon this algorithm.